refactor: adopt feature-based frontend architecture
Move screens and feature-specific components into features/{auth,
workouts,exercises,plans,stats,menu}, each exposing a barrel index.ts.
components/ now holds shared UI only (app, navigation, layouts, ui, icons).

Update CLAUDE.md project layout and docs/modules/offline-sync.md
(SyncFailureBanner now lives in components/app/) to match the new
structure.
