+++ frontend/src/components/index.ts
@@ -1,5 +1,4 @@
 export * from './icons'
 export * from './navigation'
-export * from './screens'
 export * from './layouts'
 export * from './ui'
