--- frontend/src/components/modals/index.ts
@@ -1,1 +1,0 @@
-export { WorkoutFormModal } from "./WorkoutFormModal";
