--- frontend/src/components/screens/index.ts
@@ -1,13 +1,0 @@
-export { TrainingsScreen } from "./TrainingsScreen";
-export { ExercisesScreen } from "./ExercisesScreen";
-export { AddExerciseScreen } from "./AddExerciseScreen";
-export { EditExerciseScreen } from "./EditExerciseScreen";
-export { StatsScreen } from "./StatsScreen";
-export { StatsExerciseDetailScreen } from "./StatsExerciseDetailScreen";
-export { MenuScreen } from "./MenuScreen";
-export { WorkoutDetailScreen } from "./workout-detail";
-export { ExerciseSelectionModal } from "./ExerciseSelectionModal";
-export { LoginScreen } from "./LoginScreen";
-export { RegisterScreen } from "./RegisterScreen";
-export { PlansScreen } from "./PlansScreen";
-export { PlanFormScreen } from "./PlanFormScreen";
