+++ frontend/src/features/auth/index.ts
@@ -1,0 +1,2 @@
+export { LoginScreen } from "./components/LoginScreen";
+export { RegisterScreen } from "./components/RegisterScreen";
