+++ frontend/src/features/exercises/components/ExerciseList.tsx
@@ -2,7 +2,7 @@ import { useState, useMemo } from "react";
 import { useData } from "@/contexts/data";
 import type { Exercise } from "@/types";
 import type { ExerciseStats } from "@/types";
-import { MUSCLE_GROUPS } from "../../constants";
+import { MUSCLE_GROUPS } from "@/constants";
 import { useAuth } from "@/contexts/AuthContext";
 
 interface ExerciseListProps {
