+++ frontend/src/features/exercises/index.ts
@@ -1,0 +1,5 @@
+export { ExercisesScreen } from "./components/ExercisesScreen";
+export { AddExerciseScreen } from "./components/AddExerciseScreen";
+export { EditExerciseScreen } from "./components/EditExerciseScreen";
+export { ExerciseSelectionModal } from "./components/ExerciseSelectionModal";
+export { ExerciseList } from "./components/ExerciseList";
