+++ frontend/src/features/menu/index.ts
@@ -1,0 +1,1 @@
+export { MenuScreen } from "./components/MenuScreen";
