+++ frontend/src/features/plans/index.ts
@@ -1,0 +1,2 @@
+export { PlansScreen } from "./components/PlansScreen";
+export { PlanFormScreen } from "./components/PlanFormScreen";
