+++ frontend/src/features/stats/components/StatsProgressChart.tsx
