+++ frontend/src/features/stats/components/StatsScreen.tsx
