+++ frontend/src/features/stats/index.ts
@@ -1,0 +1,2 @@
+export { StatsScreen } from "./components/StatsScreen";
+export { StatsExerciseDetailScreen } from "./components/StatsExerciseDetailScreen";
