+++ frontend/src/features/workouts/components/workout-detail/SetRowEditable.tsx
