+++ frontend/src/features/workouts/components/workout-detail/WorkoutItemCard.tsx
