+++ frontend/src/features/workouts/components/workout-detail/formatters.ts
