+++ frontend/src/features/workouts/components/workout-detail/index.ts
