+++ frontend/src/features/workouts/index.ts
@@ -1,0 +1,3 @@
+export { TrainingsScreen } from "./components/TrainingsScreen";
+export { WorkoutDetailScreen } from "./components/workout-detail";
+export { WorkoutFormModal } from "./components/WorkoutFormModal";
